+++ sphinx/source/changelog.rst
@@ -3,6 +3,7 @@ Changelog
 
 1.1.2b1
 -------
+* Added a theme tune and an 'all shields' tune for Skool Daze Take Too
 
 1.1.1 (2014-01-26)
 ------------------