+++ sphinx/source/changelog.rst
@@ -7,6 +7,7 @@ Changelog
   year' tunes in Skool Daze Take Too, Ezad Looks and Back to Skool Daze
 * Customised the lesson questions and answers in Skool Daze Take Too, Ezad
   Looks and Back to Skool Daze
+* Customised the 'sit down' messages in Back to Skool Daze
 
 1.1.1 (2014-01-26)
 ------------------